+++ bauh/api/constants.py
@@ -2,3 +2,4 @@ from pathlib import Path
 
 HOME_PATH = Path.home()
 CACHE_PATH = '{}/.cache/bauh'.format(HOME_PATH)
+CONFIG_PATH = '{}/.config/bauh'.format(HOME_PATH)